+++ packages/cli/src/services/output.ts
@@ -1,7 +1,7 @@
 import * as process from 'process';
 import { IRuleResult } from '@stoplight/spectral-core';
 import { promises as fs } from 'fs';
-import { html, json, junit, stylish, teamcity, text, pretty } from '@stoplight/spectral-formatters';
+import { html, json, junit, stylish, teamcity, text, pretty, githubActions } from '@stoplight/spectral-formatters';
 import type { Formatter, FormatterOptions } from '@stoplight/spectral-formatters';
 import type { OutputFormat } from './config';
 
@@ -13,6 +13,7 @@ const formatters: Record<OutputFormat, Formatter> = {
   html,
   text,
   teamcity,
+  'github-actions': githubActions,
 };
 
 export function formatOutput(results: IRuleResult[], format: OutputFormat, formatOptions: FormatterOptions): string {